Reset OSRELEASE in container before jlab.sh
OSRELEASE is inherited from the calling system. If OSRELEASE is defined there, it is likely not set to the correct value for the container. In that case, reset OSRELEASE and get jlab.sh to redefine it according to what's it the container.

Although one might argue that the user should unset OSRELEASE before using the container, this will likely stump users who have jlab.sh already loaded.